requirements.txt
webdriver-manager==1.7
eyes-selenium>=4.1.*